// enums/http-status-code.enum.ts
export enum HttpStatusCode {
  INTERNAL_SERVER_ERROR = '500',
  SIGNATURE_TRACKING = '039',
  INVALID_TOKEN = 498,
  SIGNATURE_TRACKING_AUTHORIZATION = '43',
  PROCESS_PAYMENT_OF_PAYROLL = '035',
  TIME_OUT = '504',
  SIGNATURE_TRACKING_MODIFY_SUCCESS = '044',
  SUCCESS_TRANSACTION = 200,
  CHANGE_PASSWORD = '104',
  SIGNATURE_TRACKING_PENDING_AUTHORIZATION = '0043'
}